# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProManageX Enterprise SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contractors:
    get:
      operationId: listContractors
      tags: [contractors]
      summary: List all contractors
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: classification
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contractors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contractor"
    post:
      operationId: createContractor
      tags: [contractors]
      summary: Create a new contractor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractorBody"
      responses:
        "201":
          description: Created contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
  /contractors/{id}:
    get:
      operationId: getContractor
      tags: [contractors]
      summary: Get a contractor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contractor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
    patch:
      operationId: updateContractor
      tags: [contractors]
      summary: Update a contractor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractorBody"
      responses:
        "200":
          description: Updated contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
    delete:
      operationId: deleteContractor
      tags: [contractors]
      summary: Delete a contractor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contractors/{id}/blacklist:
    patch:
      operationId: toggleContractorBlacklist
      tags: [contractors]
      summary: Toggle contractor blacklist status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blacklisted]
              properties:
                blacklisted:
                  type: boolean
                reason:
                  type: ["string", "null"]
      responses:
        "200":
          description: Updated contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: contractorId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a new contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: integer
        - name: workAreaId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create an attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceBody"
      responses:
        "201":
          description: Created attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendanceBody"
      responses:
        "200":
          description: Updated attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
...
  /attendance/stats:
    get:
      operationId: getAttendanceStats
      tags: [attendance]
      summary: Attendance statistics
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Attendance stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStats"

  # ─── EVALUATIONS ─────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Contractor:
      type: object
      required: [id, name, companyName, status, classification, blacklisted, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: string
        email:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        nationalId:
          type: [string, 'null']
        taxNumber:
          type: [string, 'null']
        classification:
          type: string
          enum: [tier1, tier2, tier3, preferred, blacklisted]
        status:
          type: string
          enum: [active, pending, suspended, expired]
        blacklisted:
          type: boolean
        blacklistReason:
          type: [string, 'null']
        performanceScore:
          type: ["number", "null"]
        address:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    CreateContractorBody:
      type: object
      required: [name, companyName, classification]
      properties:
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        taxNumber:
          type: string
        classification:
          type: string
        address:
          type: string
    UpdateContractorBody:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        classification:
          type: string
        status:
          type: string
        address:
          type: string
...
    Contract:
      type: object
      required: [id, title, contractorId, status, startDate, endDate, value, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        contractorId:
          type: integer
        contractorName:
          type: [string, 'null']
        status:
          type: string
          enum: [draft, active, expired, renewed, terminated]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        value:
          type: number
        paidAmount:
          type: ["number", "null"]
        description:
          type: [string, 'null']
        fileUrl:
          type: [string, 'null']
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    CreateContractBody:
      type: object
      required: [title, contractorId, startDate, endDate, value]
      properties:
        title:
          type: string
        contractorId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        value:
          type: number
        description:
          type: string
    UpdateContractBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        endDate:
          type: string
          format: date
        value:
          type: number
        paidAmount:
          type: number
        description:
          type: string
...
    Attendance:
      type: object
      required: [id, userId, date, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: [string, 'null']
        workAreaId:
          type: ["integer", "null"]
        workAreaName:
          type: [string, 'null']
        date:
          type: string
          format: date
        checkIn:
          type: [string, 'null']
          format: date-time
        checkOut:
          type: [string, 'null']
          format: date-time
        status:
          type: string
          enum: [present, absent, late, half_day, leave]
        overtimeHours:
          type: ["number", "null"]
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateAttendanceBody:
      type: object
      required: [userId, date, status]
      properties:
        userId:
          type: integer
        workAreaId:
          type: integer
        date:
          type: string
          format: date
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        status:
          type: string
        overtimeHours:
          type: number
        notes:
          type: string
    UpdateAttendanceBody:
      type: object
      properties:
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        status:
          type: string
        overtimeHours:
          type: number
        notes:
          type: string
...
    AttendanceStats:
      type: object
      required: [presentRate, absentRate, lateRate, overtimeHours, trend]
      properties:
        presentRate:
          type: number
        absentRate:
          type: number
        lateRate:
          type: number
        overtimeHours:
          type: number
        trend:
          type: array
          items:
            $ref: "#/components/schemas/TrendDataPoint"

    # ── Evaluation Schemas ──────────────────────────────────────────────────────
...
    TrendDataPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
        secondaryValue:
          type: ["number", "null"]
...